(byte)ResultFlags => (int)ResultFlags

+++ Modules/CustomRpcSender.cs
@@ -239,7 +239,7 @@ public static class CustomRpcSenderExtensions
     {
         sender.AutoStartRpc(player.NetId, (byte)RpcCalls.MurderPlayer, targetClientId)
             .WriteNetObject(target)
-            .Write((byte)ExtendedPlayerControl.ResultFlags)
+            .Write((int)ExtendedPlayerControl.ResultFlags)
             .EndRpc();
     }
 }
