+++ Patches/LadderPatch.cs
@@ -50,7 +50,7 @@ public class FallFromLadder
                       .EndRpc();
                     sender.AutoStartRpc(player.NetId, (byte)RpcCalls.MurderPlayer)
                         .WriteNetObject(player)
-                        .Write((byte)ExtendedPlayerControl.ResultFlags)
+                        .Write((int)ExtendedPlayerControl.ResultFlags)
                       .EndRpc();
                     sender.SendMessage();
                     player.NetTransform.SnapTo(targetPos);
